+++ spec/rspec/mocks/at_most_spec.rb
@@ -4,95 +4,86 @@ module RSpec
   module Mocks
     describe "at_most" do
       before(:each) do
-        @mock = RSpec::Mocks::Mock.new("test mock")
+        @double = double
       end
 
-      it "fails when at most n times method is called n plus 1 times" do
-        @mock.should_receive(:random_call).at_most(4).times
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
+      it "passes when at_most(n) is called exactly n times" do
+        @double.should_receive(:do_something).at_most(2).times
+        @double.do_something
+        @double.do_something
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(n) is called less than n times" do
+        @double.should_receive(:do_something).at_most(2).times
+        @double.do_something
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(n) is never called" do
+        @double.should_receive(:do_something).at_most(2).times
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(:once) is called once" do
+        @double.should_receive(:do_something).at_most(:once)
+        @double.do_something
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(:once) is never called" do
+        @double.should_receive(:do_something).at_most(:once)
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(:twice) is called once" do
+        @double.should_receive(:do_something).at_most(:twice)
+        @double.do_something
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(:twice) is called twice" do
+        @double.should_receive(:do_something).at_most(:twice)
+        @double.do_something
+        @double.do_something
+        @double.rspec_verify
+      end
+
+      it "passes when at_most(:twice) is never called" do
+        @double.should_receive(:do_something).at_most(:twice)
+        @double.rspec_verify
+      end
+
+      it "returns the value given by a block when at_most(:once) method is called" do
+        @double.should_receive(:to_s).at_most(:once) { "testing" }
+        @double.to_s.should eq "testing"
+        @double.rspec_verify
+      end
+
+      it "fails fast when at_most(n) times method is called n plus 1 times" do
+        @double.should_receive(:do_something).at_most(2).times
+        @double.do_something
+        @double.do_something
         lambda do
-          @mock.rspec_verify
+          @double.do_something
         end.should raise_error(RSpec::Mocks::MockExpectationError)
       end
 
-      it "fails when at most once method is called twice" do
-        @mock.should_receive(:random_call).at_most(:once)
-        @mock.random_call
-        @mock.random_call
+      it "fails fast when at_most(:once) and is called twice" do
+        @double.should_receive(:do_something).at_most(:once)
+        @double.do_something
         lambda do
-          @mock.rspec_verify
+          @double.do_something
         end.should raise_error(RSpec::Mocks::MockExpectationError)
       end
 
-      it "fails when at most twice method is called three times" do
-        @mock.should_receive(:random_call).at_most(:twice)
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
+      it "fails fast when at_most(:twice) and is called three times" do
+        @double.should_receive(:do_something).at_most(:twice)
+        @double.do_something
+        @double.do_something
         lambda do
-          @mock.rspec_verify
+          @double.do_something
         end.should raise_error(RSpec::Mocks::MockExpectationError)
-      end
-
-      it "passes when at most n times method is called exactly n times" do
-        @mock.should_receive(:random_call).at_most(4).times
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
-        @mock.rspec_verify
-      end
-
-      it "passes when at most n times method is called less than n times" do
-        @mock.should_receive(:random_call).at_most(4).times
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
-        @mock.rspec_verify
-      end
-
-      it "passes when at most n times method is never called" do
-        @mock.should_receive(:random_call).at_most(4).times
-        @mock.rspec_verify
-      end
-
-      it "passes when at most once method is called once" do
-        @mock.should_receive(:random_call).at_most(:once)
-        @mock.random_call
-        @mock.rspec_verify
-      end
-
-      it "passes when at most once method is never called" do
-        @mock.should_receive(:random_call).at_most(:once)
-        @mock.rspec_verify
-      end
-
-      it "passes when at most twice method is called once" do
-        @mock.should_receive(:random_call).at_most(:twice)
-        @mock.random_call
-        @mock.rspec_verify
-      end
-
-      it "passes when at most twice method is called twice" do
-        @mock.should_receive(:random_call).at_most(:twice)
-        @mock.random_call
-        @mock.random_call
-        @mock.rspec_verify
-      end
-
-      it "passes when at most twice method is never called" do
-        @mock.should_receive(:random_call).at_most(:twice)
-        @mock.rspec_verify
-      end
-
-      it "returns the value given by a block when the at most once method is called" do
-        @mock.should_receive(:to_s).at_most(:once) { "testing" }
-        @mock.to_s.should eq "testing"
-        @mock.rspec_verify
       end
     end
   end