+++ spec/rspec/mocks/precise_counts_spec.rb
@@ -4,54 +4,64 @@ module RSpec
   module Mocks
     describe "PreciseCounts" do
       before(:each) do
-        @mock = double("test mock")
+        @double = double("test double")
       end
 
       it "fails when exactly n times method is called less than n times" do
-        @mock.should_receive(:random_call).exactly(3).times
-        @mock.random_call
-        @mock.random_call
+        @double.should_receive(:do_something).exactly(3).times
+        @double.do_something
+        @double.do_something
         lambda do
-          @mock.rspec_verify
+          @double.rspec_verify
+        end.should raise_error(RSpec::Mocks::MockExpectationError)
+      end
+
+      it "fails fast when exactly n times method is called more than n times" do
+        @double.should_receive(:do_something).exactly(3).times
+        @double.do_something
+        @double.do_something
+        @double.do_something
+        lambda do
+          @double.do_something
         end.should raise_error(RSpec::Mocks::MockExpectationError)
       end
 
       it "fails when exactly n times method is never called" do
-        @mock.should_receive(:random_call).exactly(3).times
+        @double.should_receive(:do_something).exactly(3).times
         lambda do
-          @mock.rspec_verify
+          @double.rspec_verify
         end.should raise_error(RSpec::Mocks::MockExpectationError)
       end
 
       it "passes if exactly n times method is called exactly n times" do
-        @mock.should_receive(:random_call).exactly(3).times
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
-        @mock.rspec_verify
+        @double.should_receive(:do_something).exactly(3).times
+        @double.do_something
+        @double.do_something
+        @double.do_something
+        @double.rspec_verify
       end
 
       it "returns the value given by a block when the exactly once method is called" do
-        @mock.should_receive(:to_s).exactly(:once) { "testing" }
-        @mock.to_s.should eq "testing"
-        @mock.rspec_verify
-      end
-
-      it "passes multiple calls with different args and counts" do
-        @mock.should_receive(:random_call).twice.with(1)
-        @mock.should_receive(:random_call).once.with(2)
-        @mock.random_call(1)
-        @mock.random_call(2)
-        @mock.random_call(1)
-        @mock.rspec_verify
+        @double.should_receive(:to_s).exactly(:once) { "testing" }
+        @double.to_s.should eq "testing"
+        @double.rspec_verify
       end
 
       it "passes mutiple calls with different args" do
-        @mock.should_receive(:random_call).once.with(1)
-        @mock.should_receive(:random_call).once.with(2)
-        @mock.random_call(1)
-        @mock.random_call(2)
-        @mock.rspec_verify
+        @double.should_receive(:do_something).once.with(1)
+        @double.should_receive(:do_something).once.with(2)
+        @double.do_something(1)
+        @double.do_something(2)
+        @double.rspec_verify
+      end
+
+      it "passes multiple calls with different args and counts" do
+        @double.should_receive(:do_something).twice.with(1)
+        @double.should_receive(:do_something).once.with(2)
+        @double.do_something(1)
+        @double.do_something(2)
+        @double.do_something(1)
+        @double.rspec_verify
       end
     end
   end