+++ spec/rspec/mocks/twice_counts_spec.rb
@@ -2,65 +2,64 @@ require 'spec_helper'
 
 module RSpec
   module Mocks
-    describe "TwiceCounts" do
+    describe "#twice" do
       before(:each) do
-        @mock = double("test mock")
+        @double = double("test double")
       end
 
-      it "twice should fail when call count is higher than expected" do
-        @mock.should_receive(:random_call).twice
-        @mock.random_call
-        @mock.random_call
-        @mock.random_call
+      it "passes when called twice" do
+        @double.should_receive(:do_something).twice
+        @double.do_something
+        @double.do_something
+        @double.rspec_verify
+      end
+
+      it "passes when called twice with specified args" do
+        @double.should_receive(:do_something).twice.with("1", 1)
+        @double.do_something("1", 1)
+        @double.do_something("1", 1)
+        @double.rspec_verify
+      end
+
+      it "passes when called twice with unspecified args" do
+        @double.should_receive(:do_something).twice
+        @double.do_something("1")
+        @double.do_something(1)
+        @double.rspec_verify
+      end
+
+      it "fails fast when call count is higher than expected" do
+        @double.should_receive(:do_something).twice
+        @double.do_something
+        @double.do_something
         lambda do
-          @mock.rspec_verify
+          @double.do_something
         end.should raise_error(RSpec::Mocks::MockExpectationError)
       end
-      
-      it "twice should fail when call count is lower than expected" do
-        @mock.should_receive(:random_call).twice
-        @mock.random_call
+
+      it "fails when call count is lower than expected" do
+        @double.should_receive(:do_something).twice
+        @double.do_something
         lambda do
-          @mock.rspec_verify
+          @double.rspec_verify
         end.should raise_error(RSpec::Mocks::MockExpectationError)
       end
-      
-      it "twice should fail when called twice with wrong args on the first call" do
-        @mock.should_receive(:random_call).twice.with("1", 1)
+
+      it "fails when called wrong args on the first call" do
+        @double.should_receive(:do_something).twice.with("1", 1)
         lambda do
-          @mock.random_call(1, "1")
+          @double.do_something(1, "1")
         end.should raise_error(RSpec::Mocks::MockExpectationError)
-        @mock.rspec_reset
+        @double.rspec_reset
       end
-      
-      it "twice should fail when called twice with wrong args on the second call" do
-        @mock.should_receive(:random_call).twice.with("1", 1)
-        @mock.random_call("1", 1)
+
+      it "fails when called with wrong args on the second call" do
+        @double.should_receive(:do_something).twice.with("1", 1)
+        @double.do_something("1", 1)
         lambda do
-          @mock.random_call(1, "1")
+          @double.do_something(1, "1")
         end.should raise_error(RSpec::Mocks::MockExpectationError)
-        @mock.rspec_reset
-      end
-      
-      it "twice should pass when called twice" do
-        @mock.should_receive(:random_call).twice
-        @mock.random_call
-        @mock.random_call
-        @mock.rspec_verify
-      end
-      
-      it "twice should pass when called twice with specified args" do
-        @mock.should_receive(:random_call).twice.with("1", 1)
-        @mock.random_call("1", 1)
-        @mock.random_call("1", 1)
-        @mock.rspec_verify
-      end
-      
-      it "twice should pass when called twice with unspecified args" do
-        @mock.should_receive(:random_call).twice
-        @mock.random_call("1")
-        @mock.random_call(1)
-        @mock.rspec_verify
+        @double.rspec_reset
       end
     end
   end